Added getQuestion to StudentRestController and fixed HttpStatus in StudentRestController

+++ src/main/java/com/example/demo/restControllers/StudentRestController.java
@@ -9,6 +9,8 @@ import com.example.demo.question.QuestionService;
 import com.example.demo.uploadImages.ImageRepository;
 import com.fasterxml.jackson.annotation.JsonView;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -33,6 +35,7 @@ public class StudentRestController {
 
     @JsonView(QuestionDetails.class)
     @PostMapping("/{conceptName}/{questionText}/{questionType}/{corrected}")
+    @ResponseStatus(HttpStatus.CREATED)
     public Question newConcreteQuestion(@PathVariable String conceptName, @PathVariable String questionText, @PathVariable String questionType, @PathVariable boolean corrected){
         Question question = new Question();
         Concept concept = conceptService.findOne(conceptName);
@@ -53,31 +56,59 @@ public class StudentRestController {
         return questionService.findAll();
     }
 
+    @JsonView(QuestionDetails.class)
+    @GetMapping("/question/{id}")
+    public ResponseEntity<Question> getQuestion(@PathVariable int id){
+
+        Question question = questionService.findOne(id);
+
+        if (question != null){
+            return new ResponseEntity<>(question, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    }
+
 
     @JsonView(QuestionDetails.class)
     @RequestMapping(value = "/question/updateText/{id}/{newText}", method = RequestMethod.PUT)
-    public Question updateQuestionText(@PathVariable int id, @PathVariable String newText){
+    public ResponseEntity<Question> updateQuestionText(@PathVariable int id, @PathVariable String newText){
         Question question = questionService.findOne(id);
-        question.setQuestion(newText);
-        questionService.save(question);
-        return question;
+
+        if (question != null){
+            question.setQuestion(newText);
+            questionService.save(question);
+            return new ResponseEntity<>(question, HttpStatus.OK);
+        }else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @JsonView(QuestionDetails.class)
     @RequestMapping(value = "/question/updateCorrected/{id}/{corrected}", method = RequestMethod.PUT)
-    public Question updateQuestionCorrected(@PathVariable int id, @PathVariable boolean corrected){
+    public ResponseEntity<Question> updateQuestionCorrected(@PathVariable int id, @PathVariable boolean corrected){
         Question question = questionService.findOne(id);
-        question.setCorrected(corrected);
-        questionService.save(question);
-        return question;
+
+        if (question != null){
+            question.setCorrected(corrected);
+            questionService.save(question);
+            return new ResponseEntity<>(question, HttpStatus.OK);
+        }else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @JsonView(QuestionDetails.class)
     @RequestMapping(value = "/question/delete/{id}", method = RequestMethod.DELETE)
-    public Question deleteQuestion(@PathVariable int id){
+    public ResponseEntity<Question> deleteQuestion(@PathVariable int id){
         Question question = questionService.findOne(id);
-        questionService.delete(id);
-        return question;
+
+        if (question != null){
+            questionService.delete(id);
+            return new ResponseEntity<>(question, HttpStatus.OK);
+        }else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     // end region
@@ -86,6 +117,7 @@ public class StudentRestController {
 
     @JsonView(AnswerDetails.class)
     @PostMapping("/question/{questionId}/newAnswer/{openAnswer}/{mark}")
+    @ResponseStatus(HttpStatus.CREATED)
     public Answer newAnswer(@PathVariable int questionId, @PathVariable String openAnswer, @PathVariable boolean mark){
         Answer answer = new Answer();
         Question question = questionService.findOne(questionId);
@@ -106,38 +138,57 @@ public class StudentRestController {
 
     @JsonView(AnswerDetails.class)
     @GetMapping("/answer/{id}")
-    public Answer getAnswer(@PathVariable int id){
+    public ResponseEntity<Answer> getAnswer(@PathVariable int id){
         Answer answer = answerService.findOne(id);
-        return answer;
-    }
 
+        if(answer != null){
+            return new ResponseEntity<>(answer, HttpStatus.OK);
+        }else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    }
 
     @JsonView(AnswerDetails.class)
     @RequestMapping(value = "/answer/updateOpenAnswer/{id}/{openAnswer}", method = RequestMethod.PUT)
-    public Answer updateOpenAnswer(@PathVariable int id, @PathVariable String openAnswer){
+    public ResponseEntity<Answer> updateOpenAnswer(@PathVariable int id, @PathVariable String openAnswer){
         Answer answer = answerService.findOne(id);
-        answer.setOpenAnswer(openAnswer);
-        answerService.save(answer);
-        return answer;
+
+        if(answer != null){
+            answer.setOpenAnswer(openAnswer);
+            answerService.save(answer);
+            return new ResponseEntity<>(answer, HttpStatus.OK);
+        }else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @JsonView(AnswerDetails.class)
     @RequestMapping(value = "/answer/updateAnswerMark/{id}/{mark}", method = RequestMethod.PUT)
-    public Answer updateAnswerMark(@PathVariable int id, @PathVariable boolean mark){
+    public ResponseEntity<Answer> updateAnswerMark(@PathVariable int id, @PathVariable boolean mark){
         Answer answer = answerService.findOne(id);
-        answer.setMark(mark);
-        answerService.save(answer);
-        return answer;
+
+        if(answer != null){
+            answer.setMark(mark);
+            answerService.save(answer);
+            return new ResponseEntity<>(answer, HttpStatus.OK);
+        }else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @JsonView(AnswerDetails.class)
     @RequestMapping(value = "/question{questionId}/deleteAnswer/{id}", method = RequestMethod.DELETE)
-    public Answer deleteAnswer(@PathVariable int id, @PathVariable int questionId) {
+    public ResponseEntity<Answer> deleteAnswer(@PathVariable int id, @PathVariable int questionId) {
         Question question = questionService.findOne(questionId);
-        question.setAnswer(null);
-        Answer answer = answerService.findOne(id);
-        answerService.delete(id);
-        return answer;
+
+        if (question != null) {
+            question.setAnswer(null);
+            Answer answer = answerService.findOne(id);
+            answerService.delete(id);
+            return new ResponseEntity<>(answer, HttpStatus.OK);
+        }else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
     //end region
